refactor/fixes/use excel converter library

+++ src/main/java/com/adnanbk/ecommerce/dto/ProductDto.java
@@ -10,7 +10,6 @@ import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -33,7 +32,7 @@ public class ProductDto {
 
     @DecimalMin("0.0")
     @NotNull(message = "{error.empty}")
-    private BigDecimal unitPrice;
+    private double unitPrice;
 
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
     private List<String> images;
